+++ src/core/index/reader_slice.rs
@@ -19,12 +19,12 @@ use derive_getters::Getters;
 /// Subreader slice from a parent composite reader.
 #[derive(Debug, Clone, PartialEq, Eq, Getters, Default)]
 pub struct ReaderSlice {
-    pub start: i32,
+    pub start: usize,
     pub length: i32,
     pub reader_index: i32,
 }
 impl ReaderSlice {
-    pub fn new(start: i32, length: i32, reader_index: i32) -> Self {
+    pub fn new(start: usize, length: i32, reader_index: i32) -> Self {
         Self {
             start,
             length,
